requirements.txt
openai
requests
python-dotenv
click
rich
